Split routes from main
